convert get_battery_perct_bar to double

+++ src/common.h
@@ -60,7 +60,7 @@ void get_acpi_ac_adapter(char *, size_t);
 void get_acpi_fan(char *, size_t);
 void get_battery_stuff(char *buf, unsigned int n, const char *bat, int item);
 int get_battery_perct(const char *bat);
-uint8_t get_battery_perct_bar(struct text_object *);
+double get_battery_perct_bar(struct text_object *);
 void get_battery_short_status(char *buf, unsigned int n, const char *bat);
 
 void scan_loadavg_arg(struct text_object *, const char *);
